requirements.txt
flask
shutil
flask-wtf
imutils
opencv-python
pillow
matplotlib
numpy